refactor : final todo refactor
