+++ src/main/java/com/sandbox/domain/articles/dto/Article.java
@@ -5,7 +5,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.time.Instant;
+import java.time.LocalDateTime;
 
 @Entity
 @Getter @Setter
@@ -17,5 +17,5 @@ public class Article {
 
     private String title;
 
-    private Instant createdAt; //지금 샌드박스에 있는 시간 형식!
+    private String createdAt; //지금 샌드박스에 있는 시간 형식!
 }
